+++ annotation/src/main/java/com/blade/Demo.java
@@ -1,0 +1,4 @@
+package com.blade;
+
+public class Demo {
+}
