+++ annotation/src/main/java/com/blade/annotation/Inject.java
@@ -1,0 +1,5 @@
+package com.blade.annotation;
+
+public @interface Inject {
+    String value() default "";
+}
