+++ annotation/src/main/java/com/blade/annotation/Provides.java
@@ -1,0 +1,15 @@
+package com.blade.annotation;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD, ElementType.TYPE})
+public @interface Provides {
+    String value() default "";
+
+    boolean deepProvides() default false;
+}
