+++ annotation/src/main/java/com/blade/annotation/inter/Provider.java
@@ -1,0 +1,5 @@
+package com.blade.annotation.inter;
+
+public interface Provider {
+    Object find(String key);
+}
