--- annotation/src/main/java/com/example/annotation/Inject.java
@@ -1,5 +1,0 @@
-package com.example.annotation;
-
-public @interface Inject {
-    String value() default "";
-}
