--- annotation/src/main/java/com/example/annotation/inter/Provider.java
@@ -1,5 +1,0 @@
-package com.example.annotation.inter;
-
-public interface Provider {
-    Object find(String key);
-}
