+++ app/src/main/java/com/blade/pby/injectdemo/Demo.java
@@ -1,0 +1,8 @@
+package com.blade.pby.injectdemo;
+
+import com.blade.annotation.Provides;
+
+public class Demo {
+    @Provides(deepProvides = true, value = "demoString")
+    Demo3 string3 = new Demo3();
+}
