+++ app/src/main/java/com/blade/pby/injectdemo/demo/Demo.java
@@ -1,0 +1,8 @@
+package com.blade.pby.injectdemo.demo;
+
+import com.blade.annotation.Provides;
+
+public class Demo {
+    @Provides("demoDemoString")
+    public Object demoDemoString = "string";
+}
