--- app/src/main/java/com/example/pby/injectdemo/Demo3.java
@@ -1,17 +1,0 @@
-package com.example.pby.injectdemo;
-
-import com.example.annotation.Provides;
-import com.example.pby.injectdemo.demo.Demo;
-
-public class Demo3 {
-    @Provides(value = "demo3String")
-    String string = "demo";
-    @Provides(value = "strings")
-    String[] strings = new String[]{"string1", "string2"};
-
-    @Provides
-    String[] string1 = new String[]{"123"};
-
-    @Provides(deepProvides = true)
-    Demo demo = new Demo();
-}
