--- app/src/main/java/com/example/pby/injectdemo/MainActivity.java
@@ -1,36 +1,0 @@
-package com.example.pby.injectdemo;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.example.annotation.Inject;
-import com.example.inject.Blade;
-
-import java.util.HashMap;
-
-
-public class MainActivity extends AppCompatActivity {
-
-    @Inject("demo3String")
-    String string1;
-    @Inject("strings")
-    String[] strings;
-    @Inject
-    String[] strings1;
-    @Inject("demoDemoString")
-    String demoDemoString;
-    @Inject("int")
-    int code;
-    @Inject("pby")
-    String pby;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        Context context = new Context();
-        HashMap<String, Object> map = new HashMap<>();
-        map.put("pby", "pby123");
-        Blade.inject(this, context, map);
-    }
-}
