--- app/src/main/java/com/example/pby/injectdemo/demo/Demo.java
@@ -1,8 +1,0 @@
-package com.example.pby.injectdemo.demo;
-
-import com.example.annotation.Provides;
-
-public class Demo {
-    @Provides("demoDemoString")
-    public Object demoDemoString = "string";
-}
