--- inject/src/main/java/com/example/inject/EmptyProviderImpl.java
@@ -1,23 +1,0 @@
-package com.example.inject;
-
-import com.example.annotation.inter.Provider;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class EmptyProviderImpl implements Provider {
-
-    private Map<String, Object> pathMap;
-
-    public EmptyProviderImpl(Map<String, ?> extraMap) {
-        pathMap = new HashMap<>();
-        if (extraMap != null && !extraMap.isEmpty()) {
-            pathMap.putAll(extraMap);
-        }
-    }
-
-    @Override
-    public Object find(String key) {
-        return pathMap.get(key);
-    }
-}
