+++ processor/src/main/java/com/blade/processor/Constants.java
@@ -1,0 +1,12 @@
+package com.blade.processor;
+
+public final class Constants {
+
+    public static final String TARGET = "target";
+    public static final String SOURCE = "source";
+
+    public static final String INJECT_SUFFIX = "_Inject";
+    public static final String INJECT_METHOD_NAME = "inject";
+
+    public static final String PROVIDER_SUFFIX = "ProviderImpl";
+}
