--- processor/src/main/java/com/example/processor/Constants.java
@@ -1,12 +1,0 @@
-package com.example.processor;
-
-public final class Constants {
-
-    public static final String TARGET = "target";
-    public static final String SOURCE = "source";
-
-    public static final String INJECT_SUFFIX = "_Inject";
-    public static final String INJECT_METHOD_NAME = "inject";
-
-    public static final String PROVIDER_SUFFIX = "ProviderImpl";
-}
