--- processor/src/main/java/com/example/processor/util/StringUtils.java
@@ -1,17 +1,0 @@
-package com.example.processor.util;
-
-public class StringUtils {
-
-    private static final String EMPTY_STRING = "";
-
-    public static String emptyIfNull(String id) {
-        if (id != null) {
-            return id;
-        }
-        return EMPTY_STRING;
-    }
-
-    public static boolean isEmpty(String string) {
-        return string == null || string.equals("");
-    }
-}
